cmd: make bucket upload command take lset from flags
